app/RFJScraper/requirements.txt
appdirs==1.4.4
astroid==2.4.2
asttokens==2.0.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
bs4==0.0.1
cattrs==1.1.2
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
chromedriver-autoinstaller==0.2.2
coala-utils==0.5.1
colorama==0.4.4
cssselect==1.1.0
dateparser==1.0.0
docutils==0.16
et-xmlfile==1.0.1
facebook-scraper==0.2.17
facebook-sdk==3.1.0
fake-useragent==0.1.11
future==0.18.2
html5lib==1.1
idna==2.9
ig-profile==1.0
instagramy==4.0
isort==5.6.4
jdcal==1.4.1
jedi==0.17.2
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
MechanicalSoup==0.12.0
mechanize==0.4.5
multidict==5.1.0
mypy==0.790
mypy-extensions==0.4.3
numpy==1.19.3
oauthlib==3.1.0
openpyxl==3.0.4
pandas==1.1.5
parse==1.18.0
parso==0.7.1
Pillow==8.1.0
pycparser==2.20
pyee==7.0.4
pylint==2.6.0
PyMySQL==0.10.1
pyppeteer==0.2.3
pyquery==1.4.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-facebook-api==0.8.0
python-socks==1.1.2
python-twitter==3.5
pytz==2020.1
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
responses==0.11.0
schedule==0.6.0
Scweet==0.2.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0
thonny==3.3.1
toml==0.10.2
tqdm==4.55.1
tweepy==3.10.0
twitter-scraper==0.4.4
twitterscraper==1.6.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
